+++ src/toolbox.c
@@ -7,11 +7,11 @@
 
 #include "main.h"
 
-char *my_itoa(int nb)
+char *my_itoa(float nb)
 {
 	int len = 0;
 	char *str;
-	int tmp = nb;
+	float tmp = nb;
 
 	while (tmp > 0) {
 		tmp /= 10;
@@ -20,7 +20,7 @@ char *my_itoa(int nb)
 	str = malloc(sizeof(*str) * (len + 1));
 	str[len] = '\0';
 	while (len--) {
-		str[len] = nb % 10  + '0';
+		str[len] = (int)nb % 10  + '0';
 		nb /= 10;
 	}
 	return (str);
